pyproject.toml
[project]
name = "google-machine-learning-cc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-ml-edu>=0.1.2",
    "keras>=3.9.2",
    "matplotlib>=3.10.1",
    "numpy>=2.1.3",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "seaborn>=0.13.2",
    "tensorflow>=2.19.0",
]

[tool.uv.workspace]
members = [
    "linear_regression",
    "logistic_regression",
    "tests",
]
